+++ newrelic/resource_newrelic_synthetics_monitor.go
@@ -80,7 +80,7 @@ func resourceNewRelicSyntheticsMonitor() *schema.Resource {
 			"status": {
 				Type:         schema.TypeString,
 				Required:     true,
-				Description:  "The monitor status (i.e. ENABLED, MUTED, DISABLED). Note: The 'MUTED' status is now deprecated, and support for this value will soon be removed from the Terraform Provider in an upcoming release. It is highly recommended for users to refrain from using this value and shift to alternatives.",
+				Description:  "The monitor status (ENABLED or DISABLED).",
 				ValidateFunc: validateSyntheticMonitorStatus,
 			},
 			"validation_string": {